Update RoamingTaxaLocalisation.java
Added conditions, there is still a lot of clean up I need to do here, and will hopefully have that mostly done by the weekend.
Fixed and now correctly simplifies or doesn't simplify when it's supposed to. 
It should work for all cases, but I'll try a few more cases to make certain of that soon.